+++ clients/gtest/general_gtest.cpp
@@ -1,8 +1,8 @@
 /* ************************************************************************
  * Copyright 2020-2021 Advanced Micro Devices, Inc.
  * ************************************************************************ */
-#include "check_numerics_matrix.hpp"
-#include "check_numerics_vector.hpp"
+#include "../../library/src/include/check_numerics_matrix.hpp"
+#include "../../library/src/include/check_numerics_vector.hpp"
 #include "rocblas_data.hpp"
 #include "rocblas_test.hpp"
 #include "rocblas_vector.hpp"