+++ install.sh
@@ -17,6 +17,8 @@ rocBLAS build & installation helper script
       -i | --install             Install after build
       -d | --dependencies        Install build dependencies
       -c | --clients             Build library clients too (combines with -i & -d)
+           --clients-only        Build only clients with a pre-built library
+           --library-path        When only building clients, the path to the pre-built rocBLAS library (default is /opt/rocm/rocblas)
       -g | --debug               Set -DCMAKE_BUILD_TYPE=Debug (default is =Release)
       -f | --fork                GitHub fork to use, e.g., ROCmSoftwarePlatform or MyUserName
       -b | --branch              GitHub branch or tag to use, e.g., develop, mybranch or <commit hash>
@@ -148,6 +150,7 @@ install_msgpack_from_source( )
 # Take an array of packages as input, and delegate the work to the appropriate distro installer
 # prereq: ${ID} must be defined before calling
 # prereq: ${build_clients} must be defined before calling
+# prereq: ${tensile_msgpack_backend} must be defined before calling
 install_packages( )
 {
   if [ -z ${ID+foo} ]; then
@@ -163,7 +166,7 @@ install_packages( )
   # dependencies needed to build the rocblas library
   local library_dependencies_ubuntu=( "make" "cmake-curses-gui"
                                       "python3" "python3-yaml" "python3-venv" "python3*-pip"
-                                      "llvm-6.0-dev" "wget" "libmsgpack-dev" "libmsgpackc2" )
+                                      "wget" )
   local library_dependencies_centos_rhel=( "epel-release"
                                       "make" "cmake3" "rpm-build"
                                       "python34" "python3*-PyYAML" "python3-virtualenv"
@@ -171,21 +174,19 @@ install_packages( )
   local library_dependencies_centos_rhel_8=( "epel-release"
                                       "make" "cmake3" "rpm-build"
                                       "python3" "python3*-PyYAML" "python3-virtualenv"
-                                      "gcc-c++" "wget" "llvm-devel" "llvm-static" )
+                                      "gcc-c++" "wget" )
   local library_dependencies_fedora=( "make" "cmake" "rpm-build"
                                       "python34" "python3*-PyYAML" "python3-virtualenv"
-                                      "gcc-c++" "libcxx-devel" "wget" "llvm7.0-devel" "llvm7.0-static"
-                                      "msgpack-devel" "msgpack" )
+                                      "gcc-c++" "libcxx-devel" "wget" )
   local library_dependencies_sles=(   "make" "cmake" "python3-PyYAML" "python3-virtualenv"
-                                      "gcc-c++" "libcxxtools9" "rpm-build" "wget" "llvm7-devel" )
+                                      "gcc-c++" "libcxxtools9" "rpm-build" "wget" )
 
-  if [[ ( "${ID}" != "centos" ) || ( "${VERSION_ID}" -ge 7 ) ]]; then
-    # On CentOS-7 and greater, RPM packages for LLVM-7.0 are available. For earlier CentOS versions,
-    # we must build modern LLVM versions from src.
-    library_dependencies_centos_rhel+=( "llvm7.0-devel" "llvm7.0-static" )
+  if [[ "${tensile_msgpack_backend}" == true ]]; then
+    library_dependencies_ubuntu+=("libmsgpack-dev")
+    library_dependencies_fedora+=("msgpack-devel")
   fi
 
-  if [[ ("${ID}" == "ubuntu") && ("${VERSION_ID}" == "16.04") ]]; then
+  if [[ ("${ID}" == "ubuntu") && ("${VERSION_ID}" == "16.04") && "${tensile_msgpack_backend}" == true ]]; then
     # On Ubuntu 16.04, the version of msgpack provided in the repository is outdated, so a newer version
     # must be manually downloaded and installed.  Trying to match or exceed Ubuntu 18 default
     if ! $(dpkg -s "libmsgpackc2" &> /dev/null) || $(dpkg --compare-versions $(dpkg-query -f='${Version}' --show libmsgpackc2) lt 2.1.5-1); then
@@ -198,7 +199,9 @@ install_packages( )
 
   case "${ID}" in
     centos|rhel|sles|opensuse-leap)
-      install_msgpack_from_source
+      if [[ "${tensile_msgpack_backend}" == true ]]; then
+        install_msgpack_from_source
+      fi
       ;;
   esac
 
@@ -303,6 +306,7 @@ tensile_merge_files=
 tensile_tag=
 tensile_test_local_path=
 tensile_version=
+build_library=true
 build_clients=false
 use_cuda=false
 build_tensile=true
@@ -321,6 +325,8 @@ if ! [ -z ${ROCM_PATH+x} ]; then
     rocm_path=${ROCM_PATH}
 fi
 
+library_dir_installed=${rocm_path}/rocblas
+
 # #################################################
 # Parameter parsing
 # #################################################
@@ -328,7 +334,7 @@ fi
 # check if we have a modern version of getopt that can handle whitespace and long parameters
 getopt -T
 if [[ $? -eq 4 ]]; then
-  GETOPT_PARSE=$(getopt --name "${0}" --longoptions help,install,clients,dependencies,debug,hip-clang,no-hip-clang,merge-files,no-merge-files,no_tensile,no-tensile,tensile-host,no-tensile-host,msgpack,no-msgpack,logic:,architecture:,cov:,fork:,branch:,build_dir:,test_local_path:,cpu_ref_lib:,use-custom-version:,skipldconf,static,use-cuda,rocm-dev: --options nsrhicdgl:a:o:f:b:t:u:v: -- "$@")
+  GETOPT_PARSE=$(getopt --name "${0}" --longoptions help,install,clients,clients-only,dependencies,debug,hip-clang,no-hip-clang,merge-files,no-merge-files,no_tensile,no-tensile,tensile-host,no-tensile-host,msgpack,no-msgpack,library-path:,logic:,architecture:,cov:,fork:,branch:,build_dir:,test_local_path:,cpu_ref_lib:,use-custom-version:,skipldconf,static,use-cuda,rocm-dev: --options nsrhicdgl:a:o:f:b:t:u:v: -- "$@")
 else
   echo "Need a new version of getopt"
   exit 1
@@ -356,6 +362,13 @@ while true; do
     -c|--clients)
         build_clients=true
         shift ;;
+    --clients-only)
+        build_library=false
+        build_clients=true
+        shift ;;
+    --library-path)
+        library_dir_installed=${2}
+        shift 2 ;;
     -g|--debug)
         build_release=false
         shift ;;
@@ -613,6 +626,10 @@ pushd .
     cmake_client_options="${cmake_client_options} -DBUILD_CLIENTS_SAMPLES=ON -DBUILD_CLIENTS_TESTS=ON -DBUILD_CLIENTS_BENCHMARKS=ON -DLINK_BLIS=${LINK_BLIS} -DBUILD_DIR=${build_dir}"
   fi
 
+  if [[ "${build_library}" == false ]]; then
+    cmake_client_options="${cmake_client_options} -DSKIP_LIBRARY=ON -DROCBLAS_LIBRARY_DIR=${library_dir_installed}"
+  fi
+
   if [[ "${build_hip_clang}" == true ]]; then
       cmake_common_options="${cmake_common_options} -DRUN_HEADER_TESTING=OFF"
   fi
@@ -620,13 +637,6 @@ pushd .
   if [[ "${use_cuda}" == true ]]; then
     cmake_common_options="${cmake_common_options} -DUSE_CUDA=ON"
   fi
-
-  case "${ID}" in
-    centos|rhel)
-    cmake_common_options="${cmake_common_options} -DCMAKE_FIND_ROOT_PATH=/usr/lib64/llvm7.0/lib/cmake/"
-    ;;
-  esac
-
 
   # Uncomment for cmake debugging
   # CXX=${compiler} ${cmake_executable} -Wdev --debug-output --trace ${cmake_common_options} -DCPACK_SET_DESTDIR=OFF -DCMAKE_INSTALL_PREFIX=rocblas-install -DCPACK_PACKAGING_INSTALL_PREFIX=${rocm_path} ../..
@@ -639,7 +649,11 @@ pushd .
   fi
   check_exit_code "$?"
 
-  make -j$(nproc) install
+  if [[ "${build_library}" == true ]]; then
+    make -j$(nproc) install
+  else
+    make -j$(nproc)
+  fi
   check_exit_code "$?"
 
   # #################################################